Make backend configurable
