+++ examples/_drag-around-experimental/index.js
@@ -1,11 +1,6 @@
 /** @jsx React.DOM */
 'use strict';
 
-var ReactDND = require('react-dnd');
-ReactDND.DragDropMixin.setBackend(ReactDND.DragDropBackends.MouseMove); // FIXME: not sure about this yet
-// FIXME: this leaks out into the other examples... I have some ideas on how we can use multiple backends at the same time
-
-
 var React = require('react'),
     Container = require('./Container'),
     DragLayer = require('./DragLayer'),
