+++ modules/utils/DefaultDragSource.js
@@ -1,0 +1,18 @@
+'use strict';
+
+var invariant = require('react/lib/invariant'),
+		noop = require('./noop');
+
+var DefaultDragSource = {
+  canDrag() {
+    return true;
+  },
+
+  beginDrag() {
+    invariant(false, 'Drag source must contain a method called beginDrag. See https://github.com/gaearon/react-dnd#drag-source-api');
+  },
+
+  endDrag: noop
+};
+
+module.exports = DefaultDragSource;
